+++ nativescript-target/NsJsonpChunkTemplatePlugin.js
@@ -2,33 +2,47 @@
 	MIT License http://www.opensource.org/licenses/mit-license.php
 	Author Tobias Koppers @sokra
 */
-var ConcatSource = require("webpack-sources").ConcatSource;
-var Template = require("webpack/lib/Template");
+"use strict";
 
-function JsonpChunkTemplatePlugin() { }
+const ConcatSource = require("webpack-sources").ConcatSource;
+
+class JsonpChunkTemplatePlugin {
+	apply(chunkTemplate) {
+		chunkTemplate.hooks.render.tap(
+			"JsonpChunkTemplatePlugin",
+			(modules, chunk) => {
+				const jsonpFunction = chunkTemplate.outputOptions.jsonpFunction;
+				const globalObject = chunkTemplate.outputOptions.globalObject;
+				const source = new ConcatSource();
+				source.add(
+					`(${globalObject}[${JSON.stringify(jsonpFunction)}] = ${
+						globalObject
+					}[${JSON.stringify(jsonpFunction)}] || []).push([${JSON.stringify(
+						chunk.ids
+					)},`
+				);
+				source.add(modules);
+				const entries = [chunk.entryModule].filter(Boolean).map(m =>
+					[m.id].concat(
+						Array.from(chunk.groupsIterable)[0]
+							.chunks.filter(c => c !== chunk)
+							.map(c => c.id)
+					)
+				);
+				if (entries.length > 0) {
+					source.add(`,${JSON.stringify(entries)}`);
+				}
+				source.add("])");
+				return source;
+			}
+		);
+		chunkTemplate.hooks.hash.tap("JsonpChunkTemplatePlugin", hash => {
+			hash.update("JsonpChunkTemplatePlugin");
+			hash.update("4");
+			hash.update(`${chunkTemplate.outputOptions.jsonpFunction}`);
+			hash.update(`${chunkTemplate.outputOptions.globalObject}`);
+		});
+	}
+}
 module.exports = JsonpChunkTemplatePlugin;
 
-JsonpChunkTemplatePlugin.prototype.apply = function (chunkTemplate) {
-
-	//JSONP version
-	chunkTemplate.plugin("render", function (modules, chunk) {
-		var jsonpFunction = this.outputOptions.jsonpFunction;
-		var source = new ConcatSource();
-		source.add(jsonpFunction + "(" + JSON.stringify(chunk.ids) + ",");
-		source.add(modules);
-		var entries = [chunk.entryModule].filter(Boolean).map(function (m) {
-			return m.id;
-		});
-		if (entries.length > 0) {
-			source.add("," + JSON.stringify(entries));
-		}
-		source.add(")");
-		return source;
-	});
-	chunkTemplate.plugin("hash", function (hash) {
-		hash.update("JsonpChunkTemplatePlugin");
-		hash.update("3");
-		hash.update(this.outputOptions.jsonpFunction + "");
-		hash.update(this.outputOptions.library + "");
-	});
-};