+++ nativescript-target/index.js
@@ -1,20 +1,16 @@
 module.exports = function nativescriptTarget(compiler) {
-    var options = this;
-    var webpackLib = "webpack/lib";
+    const options = this;
+    const webpackLib = "webpack/lib";
 
-    var NsNodeGlobalsPlugin = require("./NsNodeGlobalsPlugin");
     // Custom template plugin
-    var NsJsonpTemplatePlugin = require("./NsJsonpTemplatePlugin");
+    const NsJsonpTemplatePlugin = require("./NsJsonpTemplatePlugin");
 
-    var FunctionModulePlugin = require(webpackLib + "/FunctionModulePlugin");
-    var NodeSourcePlugin = require(webpackLib + "/node/NodeSourcePlugin");
-    var LoaderTargetPlugin = require(webpackLib + "/LoaderTargetPlugin");
+    const FunctionModulePlugin = require(webpackLib + "/FunctionModulePlugin");
+    const NodeSourcePlugin = require(webpackLib + "/node/NodeSourcePlugin");
+    const LoaderTargetPlugin = require(webpackLib + "/LoaderTargetPlugin");
 
-    compiler.apply(
-        new NsNodeGlobalsPlugin(),
-        new NsJsonpTemplatePlugin(options.output),
-        new FunctionModulePlugin(options.output),
-        new NodeSourcePlugin(options.node),
-        new LoaderTargetPlugin("web")
-    );
+    new NsJsonpTemplatePlugin(options.output).apply(compiler);
+    new FunctionModulePlugin(options.output).apply(compiler);
+    new NodeSourcePlugin(options.node).apply(compiler);
+    new LoaderTargetPlugin("web").apply(compiler);
 }